Refactorizations: - Removed checks for <0 for unsigned variables - Moved variable declarations in mathlib - Removed unused functions from common - Removed spaces in front of \n in strings
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@290 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ course.h
@@ -42,7 +42,7 @@ struct TCourse {
 	string dir;
 	string author;
 	string desc[MAX_DESCRIPTION_LINES];
-	int num_lines;
+	size_t num_lines;
 	GLuint preview;
 	double width;
 	double length;
